api/requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.29
pytest==7.1.2
requests==2.25.1
